+++ src/TelegramVenue.php
@@ -135,7 +135,7 @@ class TelegramVenue extends TelegramBase implements TelegramSenderContract
     /**
      * @throws CouldNotSendNotification
      */
-    public function send(): ?ResponseInterface
+    public function send(): ResponseInterface
     {
         return $this->telegram->sendVenue($this->toArray());
     }